Adds vue.js for data bindings Redesigned the admin page Adds option in admin page to configure the number of votes for multi votes
