+++ Source/Extension.Voting/package.json
@@ -22,7 +22,10 @@
   "dependencies": {
     "jquery": "^3.3.1",
     "lodash": "^4.17.10",
-    "vss-web-extension-sdk": "^5.134.0"
+    "vss-web-extension-sdk": "^5.134.0",
+    "vue": "^2.5.17",
+    "vue-class-component": "^6.3.2",
+    "vue-property-decorator": "^7.2.0"
   },
   "devDependencies": {
     "@types/jquery": "^3.3.4",
