+++ Source/Extension.Voting/src/adminPage/adminPage.html
@@ -1,0 +1,130 @@
+﻿<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8" />
+    <title>Voting</title>
+    <script src="vss.js"></script>
+    <% htmlWebpackPlugin.files.css.forEach(function (src) { %>
+    <link rel="stylesheet" href="<%= src.replace('dist/', '') %>">
+    <% }) %>
+</head>
+
+<body>
+    <div class="hide" id="adminPage">
+        <div id="waitContainerAdmin"></div>
+        <div class="vso-container">
+            <div class="row">
+                <h1 class="votingheadline">Voting settings</h1>
+                <div class="hub-pivot">
+                    <div class="filters filter-container"></div>
+                </div>
+                <div id="progressBar" class="hub-progress pageProgressIndicator col-sm-12" style="display: none;"></div>
+            </div>
+
+            <div class="row">
+                <div id="no-permissions" class="error-message col-sm-4" v-bind:class="{ hide: userIsAdmin }">
+                    <span>You do not have the required permissions to perform any action.</span>
+                </div>
+            </div>
+
+            <div class="row">
+                <div id="menueBar-container" class="toolbar col-sm-12 hide" style="padding-right:30px"></div>
+            </div>
+
+            <div id="content" style="padding-top: 10px;" class="row container" v-bind:class="{ hide: !showContent }">
+                <div class="form-group">Changing the voting level or multiple votes per item will reset the votes.</div>
+
+                <div class="form-group">
+                    <label>Title <span class="required" v-if="actualVoting.title === ''">*</span></label>
+                    <input type="text" class="form-control" v-model="actualVoting.title">
+                </div>
+
+                <div class="form-group">
+                    <label>Description</label>
+                    <input type="text" class="form-control" name="description" v-model="actualVoting.description">
+                </div>
+
+                <div class="form-group">
+                    <label>Voting level</label>
+                    <select class="form-control" name="level" v-model="actualVoting.level">
+                        <option v-for="value in levels">{{ value }}</option>
+                    </select>
+                </div>
+
+                <div class="checkbox">
+                    <label><input type="checkbox" name="multipleVotes" v-model="actualVoting.isMultipleVotingEnabled" @change="isMultipleVotingEnabledChanged">
+                        Multiple votes per item</span></label>
+                </div>
+
+                <div class="form-group" v-if="actualVoting.isMultipleVotingEnabled">
+                    <label>Votes per item</label>
+                    <input type="text" class="form-control" name="numberOfVotes" v-model="actualVoting.numberOfVotes">
+                </div>
+            </div>
+
+            <div class="row" style="padding-top: 10px;">
+                <div class="col-sm-12" id="errorMessage" v-bind:class="{ hide: showContent }">
+                    <p id="votingdisabledmessage">There is no voting available. Please create a new voting.</p>
+                </div>
+            </div>
+        </div>
+
+        <div class="modal fade" id="excludeModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
+            <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                        <h4 class="modal-title" id="myModalLabel">Included work item types</h4>
+                    </div>
+                    <div class="modal-body">
+                        <div id="includeList" class="panel panel-default pull-left" style="width:45%">
+                            <div class="panel-heading">Included work item types</div>
+                            <div class="panel-body" id="includeListbody" v-on:drop="addToIncludes" v-on:dragover="onDragOver">
+                                <div name="itemPanel" class="panel panel-default" draggable="true" v-on:dragstart="startDrag" v-for="item in levels"
+                                    v-bind:id="item.replace(/\s+/g, '')">
+                                    <div class="panel-body">
+                                        {{ item }}
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+
+                        <div id="excludeList" class="panel panel-default pull-right" style="width:45%">
+                            <div class="panel-heading">Excluded work item types</div>
+                            <div class="panel-body" id="excludeListbody" v-on:drop="addToExcludes" v-on:dragover="onDragOver">
+                                <div name="itemPanel" class="panel panel-default" draggable="true" v-on:dragstart="startDrag" v-for="item in adminPageService.excludes"
+                                    v-bind:id="item.replace(/\s+/g, '')">
+                                    <div class="panel-body">
+                                        {{ item }}
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+
+                        <div class="clearfix"></div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <span id="appVersion" style="color:#FFF;float:right;"></span>
+    </div>
+
+    <script type="text/javascript">
+        VSS.init({
+            explicitNotifyLoaded: true,
+            usePlatformScripts: true,
+            usePlatformStyles: true,
+            moduleLoaderConfig: {}
+        });
+
+        VSS.require(["vendors.js", "admin.js"], function () {
+            VSS.notifyLoadSucceeded();
+        });
+    </script>
+</body>
+
+</html>
