+++ Source/Extension.Voting/src/entities/voting.ts
@@ -1,7 +1,7 @@
 ﻿export class Voting {
     public description = "";
     public isVotingEnabled = false;
-    public numberOfVotes: number = 1; 
+    public numberOfVotes: number = 3; 
     public isVotingPaused = false;
     public isMultipleVotingEnabled = false;
     public group = "Team";
