+++ Source/Extension.Voting/webpack.dev.config.js
@@ -15,10 +15,6 @@ module.exports = {
 	module: {
 		rules: [
 			{
-				test: /\.js$/,
-				loader: 'raw-loader'
-			},
-			{
 				test: /\.ts$/,
 				loader: 'ts-loader'
 			},
@@ -33,8 +29,8 @@ module.exports = {
 	},
 	devtool: "inline-source-map",
 	entry: {
-		voting: './src/views/votingPage.ts',
-		admin: './src/views/adminPage.ts'
+		voting: './src/votingPage/votingPage.ts',
+		admin: './src/adminPage/adminPage.ts'
 	},
 	output: {
 		filename: '[name].js',
@@ -45,7 +41,10 @@ module.exports = {
 	resolve: {
 		extensions: [".ts", ".tsx", ".js"],
 		modules: ["./node_modules"],
-		symlinks: true
+		symlinks: true,
+		alias: {
+			'vue$': 'vue/dist/vue.esm.js' // 'vue/dist/vue.common.js' for webpack 1
+		}
 	},
 	plugins: [
 		new CleanWebpackPlugin(["dist"]),
@@ -73,12 +72,12 @@ module.exports = {
 		new NoEmitOnErrorsPlugin(),
 		new ProgressPlugin(),
 		new HtmlWebpackPlugin({
-			template: "./src/views/adminPage.html",
+			template: "./src/adminPage/adminPage.html",
 			filename: "adminPage.html",
 			inject: false
 		}),
 		new HtmlWebpackPlugin({
-			template: "./src/views/votingPage.html",
+			template: "./src/votingPage/votingPage.html",
 			filename: "votingPage.html",
 			inject: false
 		}),
